now statically callable
