+++ lib/stripe.rb
@@ -37,6 +37,7 @@ require "stripe/stripe_response"
 require "stripe/list_object"
 require "stripe/error_object"
 require "stripe/api_resource"
+require "stripe/api_resource_test_helpers"
 require "stripe/singleton_api_resource"
 require "stripe/webhook"
 require "stripe/stripe_configuration"
